+++ src/actions/tester.js
@@ -1,9 +1,10 @@
-import { twitterHandleRX } from '../utils';
+import { twitterHandleRX, convertLegacyAPIResponse } from '../utils';
 
 export const SET_SCREEN_NAME = Symbol('SET_SCREEN_NAME');
 export const VALIDATE_SCREEN_NAME = Symbol('VALIDATE_SCREEN_NAME');
 export const RUN_TEST = Symbol('RUN_TEST');
 export const SET_CURRENT_RESULTS = Symbol('SET_CURRENT_RESULTS');
+export const RESET_CURRENT_RESULTS = Symbol('RESET_CURRENT_RESULTS');
 export const SET_FETCH_ERROR = Symbol('SET_FETCH_ERROR');
 
 export const symbols = {
@@ -11,6 +12,7 @@ export const symbols = {
   VALIDATE_SCREEN_NAME,
   RUN_TEST,
   SET_CURRENT_RESULTS,
+  RESET_CURRENT_RESULTS,
   SET_FETCH_ERROR
 };
 
@@ -25,23 +27,25 @@ export const actions = {
   }),
   [RUN_TEST]: (state, action) => ({
     ...state,
-    loading: true
+    loading: true,
+    fetchError: null
   }),
   [SET_CURRENT_RESULTS]: (state, action) => ({
     ...state,
     loading: false,
-    currentResult: action.result
+    currentResult: convertLegacyAPIResponse(action.result)
   }),
-  [SET_FETCH_ERROR]: (state, action) => {
-    console.log(action)
-    return ({
-      ...state,
-      loading: false,
-      currentResult: {
-        errorMessage: action.errorMessage
-      }
-    })
-  }
+  [RESET_CURRENT_RESULTS]: (state, action, initialState) => ({
+    ...state,
+    loading: false,
+    currentResult: initialState.currentResult,
+    fetchError: null
+  }),
+  [SET_FETCH_ERROR]: (state, action) => ({
+    ...state,
+    loading: false,
+    fetchError: action.fetchError
+  })
 };
 
 export default { symbols, actions };