+++ src/routes/Tester.js
@@ -8,6 +8,7 @@ import {
   VALIDATE_SCREEN_NAME,
   RUN_TEST,
   SET_CURRENT_RESULTS,
+  RESET_CURRENT_RESULTS,
   SET_FETCH_ERROR
 } from '../actions/tester';
 
@@ -19,28 +20,31 @@ import Loading from '../Loading';
 
 createStore('tester', initialState, reducer);
 
+const fetchTestResults = async (dispatch, ...args) => {
+  dispatch({ type: RESET_CURRENT_RESULTS })
+  dispatch({ type: RUN_TEST });
+  let res;
+  try {
+    res = await fetch(...args);
+    const result = await res.json();
+    dispatch({ type: SET_CURRENT_RESULTS, result });
+  } catch (err) {
+    dispatch({
+      type: SET_FETCH_ERROR,
+      fetchError: { code: 'EFETCHFAILED' }
+    });
+  }
+};
+
 const Tester = (props) => {
   const { screenName } = props;
   const [{ valid }, dispatch] = useStore('tester');
+  const fetcher = fetchTestResults.bind(null, dispatch);
+
   useSWR(
     screenName && valid ? `${process.env.REACT_APP_TEST_URL}/${screenName}` : null,
     {
-      fetcher: async (...args) => {
-        dispatch({ type: RUN_TEST });
-        let res;
-        try {
-          res = await fetch(...args);
-          const result = await res.json();
-          dispatch({ type: SET_CURRENT_RESULTS, result });
-        } catch (err) {
-          dispatch({
-            type: SET_FETCH_ERROR,
-            errorMessage: err.message === 'Failed to fetch'
-              ? 'Backend is offline! Please try again later'
-              : err.message
-          });
-        }
-      },
+      fetcher,
       revalidateOnFocus: false,
       shouldRetryOnError: false
     }
@@ -48,8 +52,13 @@ const Tester = (props) => {
 
   React.useEffect(() => {
     if (screenName) {
+      document.title = `Twitter Shadowban Test ~ ${screenName}`;
       dispatch({ type: SET_SCREEN_NAME, screenName });
       dispatch({ type: VALIDATE_SCREEN_NAME, screenName });
+    } else {
+      document.title = 'Twitter Shadowban Test';
+      dispatch({ type: SET_SCREEN_NAME, screenName: '' });
+      dispatch({ type: RESET_CURRENT_RESULTS });
     }
   }, [screenName, dispatch]);
 
@@ -60,7 +69,7 @@ const Tester = (props) => {
           <Title />
           <DonateModal />
         </div>
-        <Controls />
+        <Controls fetcher={fetcher} />
         <Results />
       </div>
     </Suspense>